third_party/WebKit/LayoutTests/fast/css/non-ASCII-case-insensitive-matching-quirks-mode.html

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>

<meta charset="UTF-8">
<style>
    div { color: red }
    .Aå { color: green }
    #Aä { color: green }
</style>
<div class="aÅ"></div>
<div class="AÅ"></div>
<div class="aå"></div>
<div class="Aå"></div>

<div id="aÄ"></div>
<div id="AÄ"></div>
<div id="aä"></div>
<div id="Aä"></div>

<script>
test(function() {
  var divs = document.getElementsByTagName("div");
  assert_equals(getComputedStyle(divs[0]).color, "rgb(255, 0, 0)");
  assert_equals(getComputedStyle(divs[1]).color, "rgb(255, 0, 0)");
  assert_equals(getComputedStyle(divs[2]).color, "rgb(0, 128, 0)");
  assert_equals(getComputedStyle(divs[3]).color, "rgb(0, 128, 0)");
  assert_equals(getComputedStyle(divs[4]).color, "rgb(255, 0, 0)");
  assert_equals(getComputedStyle(divs[5]).color, "rgb(255, 0, 0)");
  assert_equals(getComputedStyle(divs[6]).color, "rgb(0, 128, 0)");
  assert_equals(getComputedStyle(divs[7]).color, "rgb(0, 128, 0)");
}, "Ensure that case folding only happens in the ASCII range when in quirks mode");
</script>
